backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# CORS and middleware
python-multipart>=0.0.20

# External API integrations
anthropic>=0.40.0
requests>=2.32.0

# Data handling
python-dateutil>=2.9.0

# Export functionality
reportlab>=4.2.0

# Environment variables
python-dotenv>=1.0.0
